+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 0.2.2
 -----
 * Fixed `isInfinite`.
+* Added support for older GHCs. On GHC < 7.8 the pattern synonyms are disabled.
 
 0.2.1
 -----